backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
transformers==4.35.2
torch==2.1.2
numpy==1.26.2
pydantic==2.5.1
python-dotenv==1.0.0
pronouncing==0.2.0 